Update bash test

+++ src/test/bash/test.sh
@@ -1,11 +1,9 @@
 #!/bin/bash
-
-# Test without env vars
-# VARS
 mkdir -p logs
-counter=0
-relative_dir="/logs/"
-logs_dir=$PWD$relative_dir
+# Test without env vars
+cd logs
+logs_dir=$PWD
+cd ../
 SERVER_PORT=8082
 
 printf "Script location: $PWD\n"
@@ -14,15 +12,15 @@ printf "Maven dir:       $PWD\n"
 
 last_pid=$(netstat -anp | grep -i "$SERVER_PORT" | grep -oE "[0-9]+/" | grep -oE "[0-9]+")
 echo "$last_pid"
-kill "$last_pid"
+[[ ! -z "$last_pid" ]] && kill "$last_pid"
 
 echo ""
 
-{ mvn spring-boot:run | tee -a "$logs_dir""mvn-log.txt" & }
+{ mvn spring-boot:run | tee -a "$logs_dir""/mvn-log.txt" & }
 
 tee_process_pid=$!
 # 15 seconds until UP
-sleep 45
+sleep 50
 
 kill "$tee_process_pid"
 last_pid=$(netstat -anp | grep -i "$SERVER_PORT" | grep -oE "[0-9]+/" | grep -oE "[0-9]+")
